+++ objc gitlab api/utilities/GLGitlabApi+Issues.h
@@ -17,7 +17,7 @@
 
 - (GLNetworkOperation *)getAllIssuesForProjectId:(int64_t)projectId
                                     privateToken:(NSString *)privateToken
-                                            page:(int)page
+                                            page:(NSInteger)page
                                 withSuccessBlock:(GLGitlabSuccessBlock)success
                                  andFailureBlock:(GLGitlabFailureBlock)failure;
