+++ objc gitlab api/utilities/GLGitlabApi+Notes.h
@@ -28,7 +28,7 @@
 // Issue notes
 - (GLNetworkOperation *)getAllNotesForIssue:(GLIssue *)issue
                                privateToken:(NSString *)privateToken
-                                       page:(int)page
+                                       page:(NSInteger)page
                            withSuccessBlock:(GLGitlabSuccessBlock)success
                             andFailureBlock:(GLGitlabFailureBlock)failure;
